requirements.txt
anyascii==0.3.2
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
colorama==0.4.6
defusedxml==0.7.1
distlib==0.3.9
Django==5.0.9
django-environ==0.11.2
django-filter==24.3
django-modelcluster==6.3
django-permissionedforms==0.1
django-taggit==5.0.1
django-treebeard==4.7.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
draftjs-exporter==5.0.0
et_xmlfile==2.0.0
filelock==3.16.1
filetype==1.2.0
identify==2.6.1
idna==3.10
iniconfig==2.0.0
l18n==2021.3
laces==0.1.1
nodeenv==1.9.1
openpyxl==3.1.5
packaging==24.1
pillow==10.4.0
pillow_heif==0.20.0
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rest-framework-simplejwt==0.0.2
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
telepath==0.3.1
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.27.0
wagtail==6.3
wagtail-modeladmin==2.1.0
Willow==1.9.0
